clean build, subclassing codable, ui button, structure of app, and more.

+++ TimeFountain/API/TDAmeritrade/Account/Balances.swift
@@ -9,14 +9,14 @@
 import Foundation
 
 struct Balances: Codable {
-    let accruedInterest, cashBalance, cashReceipts, longOptionMarketValue: Int
-    let liquidationValue, longMarketValue, moneyMarketFund, savings: Int
-    let shortMarketValue, pendingDeposits, availableFunds, availableFundsNonMarginableTrade: Int
-    let buyingPower, buyingPowerNonMarginableTrade, dayTradingBuyingPower, dayTradingBuyingPowerCall: Int
-    let equity, equityPercentage, longMarginValue, maintenanceCall: Int
-    let maintenanceRequirement, marginBalance, regTCall, shortBalance: Int
-    let shortMarginValue, shortOptionMarketValue, sma, mutualFundValue: Int
-    let bondValue: Int
+    let accruedInterest, cashBalance, cashReceipts, longOptionMarketValue: Double
+    let liquidationValue, longMarketValue, moneyMarketFund, savings: Double
+    let shortMarketValue, pendingDeposits, availableFunds, availableFundsNonMarginableTrade: Double
+    let buyingPower, buyingPowerNonMarginableTrade, dayTradingBuyingPower, dayTradingBuyingPowerCall: Double
+    let equity, equityPercentage, longMarginValue, maintenanceCall: Double
+    let maintenanceRequirement, marginBalance, regTCall, shortBalance: Double
+    let shortMarginValue, shortOptionMarketValue, sma, mutualFundValue: Double
+    let bondValue: Double
     let isInCall: Bool
-    let stockBuyingPower, optionBuyingPower: Int
+    let stockBuyingPower, optionBuyingPower: Double
 }
